Added a temporary directory for refactoring.
I'm trying to clean the project and remove some stuff, while
also improving abstractions. General rule of thumb is that the least
a class knows about another class, and the less one tiny class knows about
the whole system, the better it is.
I advice everybody NOT to merge this shit right now. It is READ-ONLY.